requirements.txt
fastapi==0.105.0
fastapi-pagination==0.11.2
uvicorn==0.24.0.post1
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]
pymysql==1.1.0
cryptography==41.0.7
pydantic==1.10.4
databases==0.6.2
alembic==1.9.2
python-dotenv==0.21.1
boto3==1.28.2
python-multipart==0.0.9
fastapi_mail==1.3.1
jinja2==3.1.2
pytz==2024.1
websockets==11.0.3
apscheduler==3.10.4